01_extract/requirements.txt
PyMuPDF
pytesseract
PyPDF2
Pillow
psycopg2-binary
pdf2image
setuptools
paddlepaddle
paddleocr
